feat: add GIS utility functions for coastline and transect processing
